pyproject.toml
[project]
run-etl = "run_pipeline.main:main"
name = "genaidrivenetl"
version = "0.1.0"
description = ""
authors = [
    {name = "Roman Devyatov",email = "romandevyatov98@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0.0"
dependencies = [
    "fastapi (>=0.128.0,<0.129.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "openai (>=2.16.0,<3.0.0)",
    "pytest (>=9.0.2,<10.0.0)",
    "psycopg[binary] (>=3.3.2,<4.0.0)",
    "pandas (>=3.0.0,<4.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "sqlalchemy (>=2.0.46,<3.0.0)",
    "sqlparse (>=0.5.5,<0.6.0)",
    "langchain (>=1.2.10,<2.0.0)",
    "langchain-openai (>=1.1.8,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "flake8 (>=7.3.0,<8.0.0)",
    "black (>=26.1.0,<27.0.0)",
    "isort (>=7.0.0,<8.0.0)"
]

[tool.black]
line-length = 88
target-version = ["py311"]
skip-string-normalization = false

[tool.isort]
profile = "black"
line_length = 88
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
